[deploy] Abstract container engine
So the `storkctl` binary is `amd64` only, which meant I needed to use
CPU emulation via the `--arch` option in Podman.

Unfortunately, I wasted half a day trying to get this to work with my
Lima machine. Using a machine created by `podman machine` solved this,
but that feels wrong and I'm getting to the bottom of that.

After I got that to work, I ran into problems mounting the container
engine socket inside of the Ansible container created by Compose.
`podman run` seems to use macOS `realpath` to resolve directories.
macOS resolves system directories, like the `/run` directory the Podman
socket is created in, to `/private/run`, a directory that doesn't exist
within the machine running the Podman daemon.

Docker doesn't do this, so I spent time NOT trying to get the app and
database working in ACM to modify the deploy script so that it can use
Docker or Podman.

(It's insane to me that it would have been faster to just run `storkctl`
inside of a Kubernetes deployment within OpenShift, secrets handling
aside. I avoided going this direction because, unfortunately, `storkctl` uses Kubeconfigs to create cluster pair
Kubernetes resources inside of the k8s clusters being paired, and
because our managed clusters are in separate namespaces, that would have
meant shuttling Kubeconfigs into the namespace `storkctl` would run
inside of, which would've meant more Ansible post-deploy hacks.)

Signed-off-by: Carlos Nunez <cnunez@redhat.com>